+++ Elements/src/Floor.cs
@@ -107,6 +107,7 @@ namespace Elements
         public override void UpdateRepresentations()
         {
             this.Representation.SolidOperations.Clear();
+            this.Openings.ForEach(o => o.UpdateRepresentations());
             this.Representation.SolidOperations.Add(new Extrude(this.Profile, this.Thickness, Vector3.ZAxis, false));
         }