+++ Elements/src/StandardWall.cs
@@ -24,6 +24,12 @@ namespace Elements
         public double Thickness { get; set; }
 
         /// <summary>
+        /// The height of the wall.
+        /// </summary>
+        public new double Height { get; protected set; }
+
+
+        /// <summary>
         /// Construct a wall along a line.
         /// </summary>
         /// <param name="centerLine">The center line of the wall.</param>
@@ -74,7 +80,8 @@ namespace Elements
         }
 
         /// <summary>
-        /// Add an opening in the wall.
+        /// Add an Opening in the Wall. The Opening x and y sets the position relative to the Wall's Centerline.Start point
+        /// in the wall elevation coordinate system.  The x and y will position origin of the rectangular opening.
         /// </summary>
         /// <param name="width">The width of the opening.</param>
         /// <param name="height">The height of the opening.</param>
@@ -91,7 +98,8 @@ namespace Elements
         }
 
         /// <summary>
-        /// Add an opening in the wall.
+        /// Add an Opening in the Wall. The Opening x and y sets the position relative to the Wall's Centerline.Start point
+        /// in the wall elevation coordinate system.  The x and y will position the origin of the polygon opening.
         /// </summary>
         /// <param name="perimeter">The perimeter of the opening.</param>
         /// <param name="x">The distance to the origin of the perimeter opening along the center line of the wall.</param>