+++ Elements/test/AdaptiveGridTests.cs
@@ -98,5 +98,26 @@ namespace Elements.Tests
             Assert.False(adaptiveGrid.TryGetVertexIndex(new Vector3(5, 5, 1), out _));
             Assert.True(adaptiveGrid.TryGetVertexIndex(new Vector3(5, 4.9, 1), out _));
         }
+
+        [Fact]
+        public void AdaptiveGridSubtractBoxSmallDifference()
+        {
+            var adaptiveGrid = new AdaptiveGrid(new Transform());
+            var polygon = Polygon.Rectangle(new Vector3(-41, -51), new Vector3(-39, -49));
+
+            var points = new List<Vector3>();
+            points.Add(new Vector3(-40, -49.9, 1));
+            points.Add(new Vector3(-40, -49.80979, 1));
+
+            adaptiveGrid.AddFromExtrude(polygon, Vector3.ZAxis, 2, points);
+            Assert.True(adaptiveGrid.TryGetVertexIndex(new Vector3(-40, -49.9, 1), out _));
+            Assert.True(adaptiveGrid.TryGetVertexIndex(new Vector3(-40, -49.9, 2), out _));
+
+            var box = new BBox3(new Vector3(-40.2, -50.190211303259034, 0),
+                                new Vector3(-39.8, -49.809788696740966, 2));
+            adaptiveGrid.SubtractBox(box);
+            Assert.False(adaptiveGrid.TryGetVertexIndex(new Vector3(-40, -49.9, 1), out _));
+            Assert.False(adaptiveGrid.TryGetVertexIndex(new Vector3(-40, -49.9, 2), out _));
+        }
     }
 }